+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-master
     * BUGFIX      #1855 [ContactBundle]   Fixed displaying correct position when adding new person to organisation
+    * HOTFIX      #1856 [MediaBundle]     Fixed delete copyright from media
 
 * 1.1.0 (2015-12-02)
     * BUGFIX      #1849 [MediaBundle]     Fixed media-edit-overlay language changer